utils/hash_map: rename to a more specific name
