+++ consensus/cw_rating_calculator/cw_rating_calculator.h
@@ -17,7 +17,7 @@
 #include "common/trinary/flex_trit.h"
 #include "consensus/tangle/tangle.h"
 #include "utils/containers/hash/hash_int64_t_map.h"
-#include "utils/hash_maps.h"
+#include "utils/hash_indexed_map.h"
 
 #define CW_RATING_CALCULATOR_LOGGER_ID "consensus_cw_rating_calculator"
